refactor identity
