+++ actix-identity/Cargo.toml
@@ -16,13 +16,14 @@ name = "actix_identity"
 path = "src/lib.rs"
 
 [dependencies]
-actix-web = { version = "4.0.0-beta.4", default-features = false, features = ["cookies", "secure-cookies"] }
 actix-service = "2.0.0-beta.5"
+actix-web = { version = "4.0.0-beta.4", default-features = false, features = ["cookies", "secure-cookies"] }
+
 futures-util = { version = "0.3.7", default-features = false }
 serde = "1.0"
 serde_json = "1.0"
 time = "0.2.23"
 
 [dev-dependencies]
-actix-rt = "2"
 actix-http = "3.0.0-beta.4"
+actix-rt = "2"
